+++ blocklets/image-bin/src/libs/utils.js
@@ -1,0 +1,4 @@
+// eslint-disable-next-line import/prefer-default-export
+export const hasAdminPermission = (user) => {
+  return ['admin', 'owner'].includes(user?.role);
+};